Refactoring: Did some renaming. Removed references to sqlite, db, etc; especially decluttered the auth flow. Rotated client secret
