--- env.local
@@ -1,3 +1,0 @@
-# Created by Vercel CLI
-NEXT_PUBLIC_SPOTIFY_CLIENT_ID="d09d3eef989f4957bd10777e7ef28e4e"
-SPOTIFY_CLIENT_SECRET="7717292bb6e3480e9aa42565a342de6e"
