+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomeLoginPage from "@/components/login";
+import HomeLoginPage from "@/components/home-login-page";
 import Image from "next/image";
 import { Helmet } from 'react-helmet';
 import { useState, useEffect } from "react";
