+++ src/components/get_top_stats.tsx
@@ -5,12 +5,11 @@ import { topTenArtistsOneMonth } from "@/lib/utils";
 import { useState } from "react";
 
 const GetTopStats = () => {
-    const [artists, setArtists] = useState([]);
+    const [artistsData, setArtists] = useState([]);
 
     const topTenArtists = async () => {
         const res = await topTenArtistsOneMonth();
         console.log(res.status);
-
         if (res.ok) {
             const data = await res.json();
             setArtists(data);
@@ -26,9 +25,9 @@ const GetTopStats = () => {
                 <Button onClick={topTenArtists}> top ten artists for one month! </Button>
             </div>
             <div>
-                {(artists.length > 0 ) &&
+                {(artistsData.length > 0 ) &&
                     <ul>
-                        {artists.map((artist: any, index: number) => (
+                        {artistsData.map((artist: any, index: number) => (
                             <li key={index}>
                                 {artist.name}
                             </li>
