+++ src/lib/utils.ts
@@ -24,25 +24,19 @@ export async function generateRandomString(length: number) {
   return values.reduce((acc, x) => acc + pool[x % pool.length],'');
 }
 
-export async function testSqlite() {
-  const res = await fetch("/api/test-sqlite");
-  return;
-}
-
 export async function redirectToSpotifyAuth() {
   window.location.href = "/api/to-spotify-auth";
-  return;
 }
 
 export async function getBaseUrl() {
   if (process.env.NODE_ENV === 'production') {
-    return 'https://gabis-wrapped.vercel.app' //make sure not to have extra '/' at the end
+    return 'https://gabis-wrapped.vercel.app'; //make sure not to have extra '/' at the end
   }
   else if (process.env.NODE_ENV === 'development') {
-    return 'http://localhost:3000'
+    return 'http://localhost:3000';
   }
   else {
-    return 'test'
+    return 'test';
   }
 }
